wip: session and wallet connection
