+++ apps/webapp/components/layout/bitcash-auth/login-dialog.tsx
@@ -1,34 +1,26 @@
 'use client'
 import { LoginDialogContent } from '@/components/layout/bitcash-auth/login-dialog-content'
 import { RegisterDialogContent } from '@/components/layout/bitcash-auth/register-dialog-content'
-import { Button, buttonVariants } from '@/components/ui/button'
+import { Button } from '@/components/ui/button'
 import { Dialog, DialogContent, DialogTrigger } from '@/components/ui/dialog'
-import { IconDownRightArrow } from '@/components/ui/icons'
 import { useErc20Balance } from '@/hooks/use-balance'
 import { useSession } from '@/hooks/use-session'
 import { cn } from '@/lib/utils'
-import { VariantProps } from 'class-variance-authority'
+
 import { LucideWallet } from 'lucide-react'
 import Link from 'next/link'
-import { useRouter, useSearchParams } from 'next/navigation'
-import { useEffect, useState } from 'react'
+import { useState } from 'react'
 import { useToggle } from 'react-use'
 import { TestnetUSDCred } from 'smartsale-contracts'
-import { runtimeEnv } from 'smartsale-lib'
+
 import { useAccount } from 'wagmi'
 
-export function BitcashLogin({
-  defaultContent = 'login',
-  buttonLabel = 'Log In',
-  buttonClassName,
-  buttonStyle
-}: BitcashAccessProps) {
+export function BitcashLogin() {
   const [open, toggleOpen] = useToggle(false)
   const [dialogContent, setDialogContent] =
-    useState<BitcashAccessContentType>(defaultContent)
-  const { session, loginUri, newSessionId } = useSession()
+    useState<BitcashAccessContentType>('login')
+  const { session } = useSession()
   const { address } = useAccount()
-  const searchParams = useSearchParams()
   const balance = useErc20Balance({
     abi: TestnetUSDCred.abi,
     contract: TestnetUSDCred.address,
@@ -36,30 +28,8 @@ export function BitcashLogin({
     chainId: TestnetUSDCred.chainId
   })
   const isLogin = dialogContent === 'login'
-  const router = useRouter()
 
-  useEffect(() => {
-    !searchParams.has('bitcash_explorer') && isLogin && toggleOpen(false)
-  }, [session, searchParams])
-
-  useEffect(() => {
-    if (!loginUri || !open || !isLogin) return
-    // post request to parent if present
-    window.parent &&
-      window.parent.postMessage({ eventType: 'esr', code: loginUri }, '*')
-
-    // redirect with esr and callback on mobile
-    if (runtimeEnv.isMobile && !searchParams.has('bitcash_explorer')) {
-      const params = new URLSearchParams()
-      params.append('esr_code', loginUri.replace('esr://', ''))
-      const callbackUrl = `${window.location.href}?session_id=${newSessionId}`
-      console.log('💥 callbackUrl', callbackUrl)
-      const encodedCallbackUrl = encodeURIComponent(callbackUrl)
-      params.append('callback', encodedCallbackUrl)
-      router.push(`https://app.bitcash.org/login?${params.toString()}`)
-    }
-  }, [open, loginUri, searchParams])
-
+  // TODO: review this, do we need it?
   if (session && isLogin)
     return (
       <Link href="#">
@@ -69,33 +39,16 @@ export function BitcashLogin({
       </Link>
     )
 
-  // never show the qr on mobile or bitcash explorer
-  const hideQr = runtimeEnv.isMobile || searchParams.has('bitcash_explorer')
-
   return (
     <Dialog open={open} onOpenChange={toggleOpen}>
       <DialogTrigger asChild>
-        <Button
-          className={cn(
-            {
-              'text-md group relative size-14 rounded-full px-0 py-0 font-bold hover:[&svg]:fill-card':
-                buttonLabel === 'down-right-icon'
-            },
-            buttonClassName
-          )}
-          {...buttonStyle}
-        >
-          {buttonLabel === 'down-right-icon' ? (
-            <IconDownRightArrow className="size-4 transition-all group-focus-within:-rotate-45 group-hover:-rotate-45 [&_path]:stroke-white" />
-          ) : (
-            buttonLabel
-          )}
+        <Button variant="secondary" radius="full" className={cn('px-10')}>
+          {session ? 'Connect' : 'Login'}
         </Button>
       </DialogTrigger>
 
-      {/* @ts-ignore */}
       <DialogContent className="box-content w-full sm:max-w-[430px]">
-        {!hideQr && isLogin && (
+        {!isLogin && (
           <LoginDialogContent updateDialogContent={setDialogContent} />
         )}
         {!isLogin && (
@@ -107,10 +60,7 @@ export function BitcashLogin({
 }
 
 interface BitcashAccessProps {
-  buttonClassName?: string
   defaultContent?: BitcashAccessContentType
-  buttonLabel?: string
-  buttonStyle?: VariantProps<typeof buttonVariants>
 }
 
 export type BitcashAccessContentType = 'login' | 'register'
