--- apps/webapp/components/layout/connect-wallet-buttons.tsx
@@ -1,37 +1,0 @@
-'use client'
-
-import { useSession } from '@/hooks/use-session'
-import { BitcashLogin } from './bitcash-auth/login-dialog'
-import { cn } from '@/lib/utils'
-import { Button, buttonVariants } from '../ui/button'
-import { useConnectModal } from '@rainbow-me/rainbowkit'
-
-export function ConnectWalletButtons() {
-  const { session } = useSession()
-  const { openConnectModal } = useConnectModal()
-  return (
-    <>
-      {session ? null : (
-        // <Button
-        //   className={cn(
-        //     buttonVariants({
-        //       variant: 'outline',
-        //       radius: 'full'
-        //     }),
-        //     'border-transparent px-10 md:border-accent'
-        //   )}
-        //   onClick={() => openConnectModal && openConnectModal()}
-        // >
-        //   Connect
-        // </Button>
-        <BitcashLogin
-          buttonStyle={{
-            variant: 'secondary',
-            radius: 'full'
-          }}
-          buttonClassName="px-10"
-        />
-      )}
-    </>
-  )
-}
